Create Random password Generator using GUI Tkinter
Password generator is a application which is designed to make strong , random and secure passwords for users use. Also taking into consideration brute force attack, it is a method used by hackers to guess a password by trying every possible combination until the 
correct one is found.
The methods which we have used is like we have used strength as low, medium, and strong. in low strength, we have used lowercase characters from a-z and numbers from 0-9. in medium strength, we have used uppercase characters from a-z, lowercase characters from a-z, and numbers from 0-9. in strong strength, we have used uppercase characters, lowercase characters, numbers, and special characters like exclamation mark, comma, brackets, hashtags, etc.
We have used two buttons in it. the first is the generate button. by clicking on the generate button, it triggers the code to create the password and display it on the password field. the second is the copy button. by clicking on the copy button, it allows the users to copy the password to the clipboard. and the password can be easily pasted on any other forms, etc.